pyproject.toml
[project]
name = "agro-platform"
version = "0.1.0"
description = "Платформа управления агрохолдингом — Modular Monolith"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.3",
    "fastapi-users[sqlalchemy]>=15.0.5",
    "fastapi-users-db-sqlalchemy>=7.0.0",
    "pydantic-settings>=2.13.1",
    "pymssql>=2.3.13",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.37",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "httpx>=0.28.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.8",
]

# ==================== Ruff Configuration ====================
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "RUF", "ARG"]
ignore = [
    "E501",      # line too long
    "RUF001",    # ← Отключаем проверку на "ambiguous" кириллицу
    "ARG001"     # ← Отключаем предупреждение о неиспользуемом аргументе `app` в lifespan
]

[tool.ruff.lint.per-file-ignores]
"src/main.py" = ["ARG001"]

# ==================== Pyright Configuration ====================
[tool.pyright]
include = ["src"]
pythonVersion = "3.12"
reportMissingImports = true